+++ NEWS.md
@@ -6,6 +6,8 @@
   hardcoded default
 - Avoid error when include_area = TRUE and polygon does not intersect raster
 - Support terra SpatRaster inputs to coverage_fraction
+- Produce error instead of incorrect result when processing SpatRaster with
+  duplicate names
 
 # version 0.7.2