+++ NEWS.md
@@ -7,6 +7,9 @@
 - Avoid error when include_area = TRUE and polygon does not intersect raster
 - Produce error instead of incorrect result when processing SpatRaster with
   duplicate names
+- Add coverage_area argument to exact_resample
+- Allow use of user-defined summary function in exact_resample
+- Allow multi-layer inputs to exact_resample
 
 # version 0.7.2